Base functionality of the image manager now in! At last!

+++ administrator/components/com_helloworld/controllers/images.php
@@ -53,11 +53,26 @@ class HelloWorldControllerImages extends JControllerForm
     // Get the image file ID of which we need to delete
     $file_id = JRequest::getVar ('file_id','','GET','int');
     
-    // Let's delete this puppy...
+    // Get the caption, needs filtering
+    $raw = JRequest::getVar ('jform','','POST','array');
+    
+    // Let's update this puppy...first we need to get the 
+    JTable::addIncludePath(JPATH_COMPONENT_ADMINISTRATOR.'/tables');	
+    $table = JTable::getInstance('Images', 'HelloWorldTable');   
+    
+    $table->id = $file_id;
+    
+		// Bind the translated fields to the JTable instance	
+		if (!$table->bind($raw))
+		{
+			JError::raiseWarning(500, $table->getError());
+		}	else {
+     $table->store($file_id);
+     $app->enqueueMessage(JText::_('COM_HELLOWORLD_IMAGES_CAPTION_SUCCESSFULLY_UPDATED'), 'message');
+    }
     
     
-    $app = JFactory::getApplication();
-    $app->enqueueMessage(JText::_('COM_HELLOWORLD_IMAGES_CAPTION_SUCCESSFULLY_UPDATED'), 'message');
+    
     
     // Set the redirection once the delete has completed...
     $this->setRedirect(JRoute::_('index.php?option=com_helloworld&task=images.edit' . $this->getRedirectToItemAppend($id, 'id'), false));
@@ -89,36 +104,44 @@ class HelloWorldControllerImages extends JControllerForm
     $table = JTable::getInstance('Images', 'HelloWorldTable');
     
     // Get the image details
-    $table->load($file_id);
-    
-    // Name of the image to remove
-    $file = $table->image_file_name;
+    if($table->load($file_id)){
+          
+      // Name of the image to remove
+      $file = $table->image_file_name;
 
-    if ($file !== JFile::makeSafe($file))
-		{
-      // filename is not safe
-			$filename = htmlspecialchars($path, ENT_COMPAT, 'UTF-8');
-			Error::raiseWarning(100, JText::sprintf('COM_MEDIA_ERROR_UNABLE_TO_DELETE_FILE_WARNFILENAME', substr($filename, strlen(COM_IMAGE_BASE))));
-    }
-    
-    $fullPaths = array();
-    // Create a path to delete the image and each of the profile images that would've been created
-  	$fullPaths[] = JPath::clean(implode(DS, array(COM_IMAGE_BASE, $id, $file)));
- 	  $fullPaths[] = JPath::clean(implode(DS, array(COM_IMAGE_BASE, $id, 'gallery', $file)));
- 	  $fullPaths[] = JPath::clean(implode(DS, array(COM_IMAGE_BASE, $id, 'thumbs', $file)));
- 	  $fullPaths[] = JPath::clean(implode(DS, array(COM_IMAGE_BASE, $id, 'thumb', $file)));
-
-    // Loop over each file path
-    foreach ($fullPaths as $path) {
-      if (is_file($path))
+      if ($file !== JFile::makeSafe($file))
       {
-        JFile::delete($path);
-        $app = JFactory::getApplication();
-      }        
-    }
-    // Set the message
-    $app->enqueueMessage(JText::_('COM_HELLOWORLD_IMAGES_IMAGE_SUCCESSFULLY_DELETED'), 'message');
+        // filename is not safe
+        $filename = htmlspecialchars($path, ENT_COMPAT, 'UTF-8');
+        Error::raiseWarning(100, JText::sprintf('COM_MEDIA_ERROR_UNABLE_TO_DELETE_FILE_WARNFILENAME', substr($filename, strlen(COM_IMAGE_BASE))));
+      }
+
+      $fullPaths = array();
+      // Create a path to delete the image and each of the profile images that would've been created
+      $fullPaths[] = JPath::clean(implode(DS, array(COM_IMAGE_BASE, $id, $file)));
+      $fullPaths[] = JPath::clean(implode(DS, array(COM_IMAGE_BASE, $id, 'gallery', $file)));
+      $fullPaths[] = JPath::clean(implode(DS, array(COM_IMAGE_BASE, $id, 'thumbs', $file)));
+      $fullPaths[] = JPath::clean(implode(DS, array(COM_IMAGE_BASE, $id, 'thumb', $file)));
+
+      // Loop over each file path
+      foreach ($fullPaths as $path) {
+        if (is_file($path))
+        {
+          JFile::delete($path);
+          $app = JFactory::getApplication();
+        }        
+      }
+      
+      $del = $table->delete($file_id);
+      
+      // Also need to check and delete this from the gallery_images table if 
    
+      
+      // Set the message
+      $app->enqueueMessage(JText::_('COM_HELLOWORLD_IMAGES_IMAGE_SUCCESSFULLY_DELETED'), 'message');
+    } else {
+      $app->enqueueMessage(JText::_('COM_HELLOWORLD_IMAGES_IMAGE_DELETE_PROBLEM_FETCHING_IMAGE_DETAILS'), 'message');
+    }
 
         
  
@@ -223,16 +246,11 @@ class HelloWorldControllerImages extends JControllerForm
         // Create a new JObject to 
         $object_file = new JObject($file);
 
+        // Move the file from the tmp location to the property image folder
         if (!JFile::upload($file['tmp_name'], $file['filepath']))
         {
           // Error in upload
           $file['error'][] = JText::_('COM_MEDIA_ERROR_UNABLE_TO_UPLOAD_FILE');
-        }
-        else
-        {
-          // Trigger the onContentAfterSave event.
-          // Should trigger this after the files have been done into the database? e.g. post process uploaded files
-          $dispatcher->trigger('onContentAfterSave', array('com_media.file', &$object_file, true));
         }
         
         // Add the url to the uploaded files array
@@ -246,23 +264,24 @@ class HelloWorldControllerImages extends JControllerForm
         $file['image_file_name'] = $file['name'];
       }
     } 
-    
-
-    // Load the relevant table model(s) so we can save the data back to the db
-    $images_model = $this->getModel('images');
-    
-    // Get an instance of the images table
-    $table = $images_model->getImagesTable(); 
-    
-    // Lastly, loop over the $files array (again) and add any new images to the existing ones
+        
+    // Lastly, loop over the $files array (again) 
     foreach ($files as &$file) {
+      
       // If the image uploaded correctly there won't be any errors
       if (count($file['error']) == 0) {
         
+        // Add a success message to the file item and then add this to an images array 
         $file['message'] = JText::_('COM_HELLOWORLD_IMAGES_IMAGE_SUCCESSFULLY_UPLOADED');
         $images[] = $file;
       } 
     }
+    
+    // Load the relevant model(s) so we can save the data back to the db
+    $images_model = $this->getModel('images');
+    
+    // Get an instance of the images table
+    $table = $images_model->getImagesTable(); 
     
     // Save the file details back to the database.
     // Need to ensure that the images are always stored against the parent property ID if this is a leaf node
@@ -271,16 +290,22 @@ class HelloWorldControllerImages extends JControllerForm
       {
         // TODO: This won't return to the user
         JError::raiseWarning(500, $table->getError());
-        //return false;
-      }      
+        $general_error[]['error'][] = JText::_('COM_HELLOWORLD_IMAGES_PROBLEM_SAVING_IMAGE_TO_DATABASE');
+      } else {
+        // Images were uploaded okay, and saved so let's create the profile images.
+        $images_model->generateImageProfiles($images, $id);
+      }     
     } else {
-      
       // Store the image against the parent property
       if (!$table->save_images($parent_id, $images))
       {
         // TODO: This won't return to the user
         JError::raiseWarning(500, $table->getError());
-        //return false;
+        $general_error[]['error'][] = JText::_('COM_HELLOWORLD_IMAGES_PROBLEM_SAVING_IMAGE_TO_DATABASE');        
+      } else {
+        // Images were uploaded okay, and saved so let's create the profile images.
+        // Trigger the onContentAfterSave event.
+        $images_model->generateImageProfiles($images, $parent_id);
       }
     }
